# server/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "OOP Labs API",
      "version": "1.0.0",
      "description": "API for managing programming lab assignments"
    },
    "servers": [
      {
        "url": "http://localhost:5000",
        "description": "Local development server"
      },
      {
        "url": "https://your-production-url.com",
        "description": "Production server"
      }
    ],
    "tags": [
      {
        "name": "Assignments",
        "description": "Lab assignment operations"
      }
    ],
    "paths": {
      "/api/health": {
        "get": {
          "tags": ["System"],
          "summary": "Check API health",
          "responses": {
            "200": {
              "description": "API is healthy",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HealthCheck"
                  }
                }
              }
            }
          }
        }
      },
      "/api/assignments": {
        "get": {
          "tags": ["Assignments"],
          "summary": "Get all assignments",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "description": "Page number",
              "schema": {
                "type": "integer",
                "default": 1
              }
            },
            {
              "name": "limit",
              "in": "query",
              "description": "Items per page",
              "schema": {
                "type": "integer",
                "default": 10
              }
            }
          ],
          "responses": {
            "200": {
              "description": "List of assignments",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/AssignmentList"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": ["Assignments"],
          "summary": "Create new assignment",
          "security": [{"ApiKeyAuth": []}],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentInput"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Assignment created",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Assignment"
                  }
                }
              }
            }
          }
        }
      },
      "/api/assignments/{id}": {
        "delete": {
          "tags": ["Assignments"],
          "summary": "Delete an assignment",
          "security": [{"ApiKeyAuth": []}],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "Assignment ID",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Assignment deleted",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SuccessResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "securitySchemes": {
        "ApiKeyAuth": {
          "type": "apiKey",
          "in": "header",
          "name": "X-Auth-Password"
        }
      },
      "schemas": {
        "HealthCheck": {
          "type": "object",
          "properties": {
            "status": {
              "type": "string",
              "example": "OK"
            },
            "timestamp": {
              "type": "string",
              "format": "date-time"
            },
            "dbStatus": {
              "type": "string",
              "example": "Connected"
            }
          }
        },
        "AssignmentInput": {
          "type": "object",
          "required": ["title", "icon", "problems"],
          "properties": {
            "title": {
              "type": "string",
              "example": "Basic OOP Concepts"
            },
            "icon": {
              "type": "string",
              "example": "FiCode"
            },
            "problems": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ProblemInput"
              }
            }
          }
        },
        "ProblemInput": {
          "type": "object",
          "required": ["question", "code", "output"],
          "properties": {
            "question": {
              "type": "string",
              "example": "Create a class named 'Car'"
            },
            "code": {
              "type": "string",
              "example": "class Car {\n  // Your code here\n}"
            },
            "output": {
              "type": "string",
              "example": "Car object created successfully"
            }
          }
        },
        "Assignment": {
          "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "example": "64a1b5e5f7d9c0123456789a"
            },
            "title": {
              "type": "string",
              "example": "Basic OOP Concepts"
            },
            "icon": {
              "type": "string",
              "example": "FiCode"
            },
            "problems": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Problem"
              }
            },
            "createdAt": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "Problem": {
          "type": "object",
          "properties": {
            "question": {
              "type": "string"
            },
            "code": {
              "type": "string"
            },
            "output": {
              "type": "string"
            }
          }
        },
        "AssignmentList": {
          "type": "object",
          "properties": {
            "success": {
              "type": "boolean"
            },
            "data": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Assignment"
              }
            },
            "pagination": {
              "$ref": "#/components/schemas/Pagination"
            }
          }
        },
        "Pagination": {
          "type": "object",
          "properties": {
            "currentPage": {
              "type": "integer"
            },
            "totalPages": {
              "type": "integer"
            },
            "totalItems": {
              "type": "integer"
            }
          }
        },
        "SuccessResponse": {
          "type": "object",
          "properties": {
            "success": {
              "type": "boolean"
            },
            "message": {
              "type": "string"
            },
            "deletedId": {
              "type": "string"
            }
          }
        }
      }
    }
  }